+++ daemons/sentry/sentry.py
@@ -32,17 +32,16 @@ from sys import exit
 from django.utils import timezone
 from ldsi.settings import TIME_ZONE
 from siem.models import LogEvent, RuleEvent, LimitRule
-#import signal
 
 
 class SiemSentry:
 
     def __init__(self, rule):
         """Initialize trigger object"""
-        
         self.rule = rule
         self.tzone = TIME_ZONE
         self.lasteventid = None
+        self.justfired = False
 
 
     def get_first_logevent(self):
@@ -130,8 +129,11 @@ class SiemSentry:
                 if self.rule.rule_events:
                     self.get_last_ruleevent()
                     expectrule = True
-            # Wait until the next interval
-            sleep(int(self.rule.time_int) * 60)
+            # Wait until next interval if firedr,; otherwise ~60 seconds:
+            if self.justfired:
+                sleep(int(self.rule.time_int) * 60)
+            else:
+                sleep(randrange(50, 70))
 
 
     def check_logevent(self):
@@ -177,6 +179,7 @@ class SiemSentry:
         
         if len(e) == 0:
             self.get_last_logevent()
+            self.justfired = False
         else:
             totalevents = sum([x.aggregated_events for x in e])
             numhosts = len({x.log_source for x in e})
@@ -210,6 +213,9 @@ class SiemSentry:
                 event.source_ids_log.set(list(e))
                 event.save()
                 self.lasteventid = e.latest('id').id
+                self.justfired = True
+            else:
+                self.justfired = False
 
 
     def check_ruleevent(self):
@@ -237,6 +243,7 @@ class SiemSentry:
 
         if len(e) == 0:
             self.get_last_ruleevent()
+            self.justfired = False
         else:
             if len(e) > self.rule.event_limit:
                 event = RuleEvent()
@@ -265,6 +272,9 @@ class SiemSentry:
                 event.source_ids_rule.set(list(e))
                 event.save()
                 self.lasteventid = e.latest('id').id
+                self.justfired = True
+            else:
+                self.justfired = False
 
 
 def start_rule(rule):